# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/feedback": {
      "get": {
        "operationId": "FeedbackController_findAll",
        "summary": "Get all feedback",
        "parameters": [
          {
            "name": "patientId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "Open",
                "In Progress",
                "Resolved"
              ],
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of feedback"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "FeedbackController_create",
        "summary": "Submit new feedback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback submission result (check success field)"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/stats/overview": {
      "get": {
        "operationId": "FeedbackController_getStats",
        "summary": "Get feedback statistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Feedback statistics retrieved"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/patient/{patientId}": {
      "get": {
        "operationId": "FeedbackController_findByPatient",
        "summary": "Get feedback by patient ID",
        "parameters": [
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient feedback retrieved"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/category/{category}": {
      "get": {
        "operationId": "FeedbackController_findByCategory",
        "summary": "Get feedback by category",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback retrieved by category"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/rating/{rating}": {
      "get": {
        "operationId": "FeedbackController_findByRating",
        "summary": "Get feedback by rating (1-5)",
        "parameters": [
          {
            "name": "rating",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback retrieved by rating"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/unresolved": {
      "get": {
        "operationId": "FeedbackController_getUnresolvedFeedback",
        "summary": "Get all unresolved feedback",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Unresolved feedback retrieved"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/high-priority": {
      "get": {
        "operationId": "FeedbackController_getHighPriorityFeedback",
        "summary": "Get high priority feedback (low ratings)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "High priority feedback retrieved"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/{id}": {
      "get": {
        "operationId": "FeedbackController_findById",
        "summary": "Get feedback by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback details retrieved"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "put": {
        "operationId": "FeedbackController_update",
        "summary": "Update feedback details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback updated successfully"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "FeedbackController_patchUpdate",
        "summary": "Partially update feedback details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback updated successfully"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "FeedbackController_delete",
        "summary": "Delete feedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback deleted successfully"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/feedback/{id}/status": {
      "patch": {
        "operationId": "FeedbackController_updateStatus",
        "summary": "Update feedback status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback status updated successfully"
          }
        },
        "tags": [
          "Feedback"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "NexCare Hospital Management API",
    "description": "Complete REST API for the NexCare Hospital Administrative Operations Platform. Features role-based access control (RBAC), JWT authentication, and comprehensive modules for patients, appointments, billing, beds, ambulance, inventory, feedback, and system management.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      },
      "x-user-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "User role for RBAC (superuser, administrative_staff, patient, ambulance, doctor, nurse)"
      }
    },
    "schemas": {
...
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "example": "P001",
            "description": "Patient ID"
          },
          "type": {
            "type": "string",
            "example": "Patient",
            "description": "Type of feedback provider"
          },
          "category": {
            "type": "string",
            "example": "service",
            "description": "Category of feedback"
          },
          "subject": {
            "type": "string",
            "example": "Great doctors",
            "description": "Subject of feedback"
          },
          "summary": {
            "type": "string",
            "example": "Dr. Smith was incredibly thorough.",
            "description": "Detailed summary"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "description": "Rating (1-5)"
          }
        },
        "required": [
          "patientId",
          "type",
          "category",
          "subject",
          "summary",
          "rating"
        ]
      },
      "UpdateFeedbackDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "example": "service",
            "description": "Category of feedback"
          },
          "subject": {
            "type": "string",
            "example": "Great doctors",
            "description": "Subject of feedback"
          },
          "summary": {
            "type": "string",
            "example": "Dr. Smith was incredibly thorough.",
            "description": "Detailed summary"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "description": "Rating (1-5)"
          },
          "status": {
            "type": "string",
            "enum": [
              "Open",
              "In Progress",
              "Resolved"
            ],
            "example": "In Progress",
            "description": "Status of feedback"
          }
        }
      },
...
    }
  }
}